Create MenuScreen.java
Bad because not implemented great and has no night screen. If I knew how to stop the actual game from starting it would be better. Not hard to implement by any means but it was the least I could do in all my stupor. In main just put  "new MenuScreen();" to start this up and see how it looks. made it simple because how pretty do we gotta make it.